requirements.txt
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.10
python-dotenv==1.1.1
